+++ src/pages/PreCadastro/PreCadastroProfissional/styles.js
@@ -1,8 +1,8 @@
-import { Paragraph } from 'react-native-paper';
 import styled from 'styled-components/native';
+import { Paragraph } from 'react-native-paper';
 import { CORES } from '~/constantes/estiloBase';
 
-// TODO: possivel remoção
+
 export const Title = styled(Paragraph)`
   font-weight: 500;
   font-size: 20px;
