--- src/pages/PreCadastro/PreCadastroSenha.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const PreCadastroSenha = () => <></>;
-
-export default PreCadastroSenha;
